Create Exercise_24
(Simulation: coupon collector’s problem) Coupon collector is a classic statistics
problem with many practical applications. The problem is to repeatedly pick
objects from a set of objects and find out how many picks are needed for all the
objects to be picked at least once. A variation of the problem is to pick cards from
a shuffled deck of 52 cards repeatedly, and find out how many picks are needed
before you see one of each suit. Assume a picked card is placed back in the deck 
before picking another. Write a program to simulate the number of picks needed
to get four cards from each suit and display the four cards picked (it is possible a
card may be picked twice). Here is a sample run of the program: