Create Exercise_04
(Compute the weekly hours for each employee) Suppose the weekly hours for all employees are stored in a two-dimensional array. Each row records an employ- ee’s seven-day work hours with seven columns. For example, the following array stores the work hours for eight employees. Write a program that displays employ- ees and their total hours in decreasing order of the total hours.