Create Exercise_13
(Locate the largest element) Write the following method that returns the location
of the largest element in a two-dimensional array:
public static int[] locateLargest(double[][] a)
 The return value is a one-dimensional array that contains two elements. These
two elements indicate the row and column indices of the largest element in the
two-dimensional array. If there are more than one largest element, return the
smallest row index and then the smallest column index.
 Write a test program that prompts the user to enter a two-dimensional array and
displays the location of the largest element in the array.