Create Exercise_25
(Markov matrix) An n * n matrix is called a positive Markov matrix if each
element is positive and the sum of the elements in each column is 1. Write the
following method to check whether a matrix is a Markov matrix:
public static boolean isMarkovMatrix(double[][] m)
 Write a test program that prompts the user to enter a 3 * 3 matrix of double
values and tests whether it is a Markov matrix. Here are sample runs:
Enter a 3−by−3 matrix row by row:
0.15 0.875 0.375
0.55 0.005 0.225
0.30 0.12 0.4
It is a Markov matrix