Create Exercise_27
(Column sorting) Implement the following method to sort the columns in a
two-dimensional array. A new array is returned and the original array is intact.
public static double[][] sortColumns(double[][] m)
Write a test program that prompts the user to enter a 3 * 3 matrix of double
values and displays a new column-sorted matrix. Here is a sample run:
Enter a 3−by−3 matrix row by row:
0.15 0.875 0.375
0.55 0.005 0.225
0.30 0.12 0.4
The column−sorted array is
0.15 0.0050 0.225
0.3 0.12 0.375
0.55 0.875 0.4