Create Exercise_31
(Geometry: intersecting point) Write a method that returns the intersecting point of
two lines. The intersecting point of the two lines can be found by using the formula
given in Programming Exercise 3.25. Assume that (x1, y1) and (x2, y2) are the
two points on line 1 and (x3, y3) and (x4, y4) are on line 2. The method header is:
public static double[] getIntersectingPoint(double[][] points)
 The points are stored in a 4-by-2 two-dimensional array points with (points
[0][0], points[0][1]) for (x1, y1). The method returns the intersecting point
or null if the two lines are parallel. Write a program that prompts the user to enter
four points and displays the intersecting point.