Create Exercise_35
(Largest block) Given a square matrix with the elements 0 or 1, write a program to
find a maximum square submatrix whose elements are all 1s. Your program should
prompt the user to enter the number of rows in the matrix. The program then displays the location of the first element in the maximum square submatrix and the
number of rows in the submatrix. 
Your program should implement and use the following method to find the maximum square submatrix:
public static int[] findLargestBlock(int[][] m)
 Here is a sample run:
Enter the number of rows in the matrix: 5
Enter the matrix row by row:
1 0 1 0 1
1 1 1 0 1
1 0 1 1 1
1 0 1 1 1
1 0 1 1 1
The maximum square submatrix is at (2, 2) with size 3