Create Exercise 04
(The MyPoint class) Design a class named MyPoint to represent a point with
x- and y-coordinates. The class contains:
■■ The data fields x and y that represent the coordinates with getter methods.
■■ A no-arg constructor that creates a point (0, 0).
■■ A constructor that constructs a point with specified coordinates.
■■ A method named distance that returns the distance from this point to a specified point of the MyPoint type.
A method named distance that returns the distance from this point to another point with specified x- and y-coordinates.
■■ A static method named distance that returns the distance from two MyPoint objects.
Write a test program that creates the two points and displays the distance between them.