Create Exercise 25
(New string split method) The split method in the String class returns an array of strings consisting of the substrings split by the delimiters. However, the
delimiters are not returned. Implement the following new method that returns an array of strings consisting of the substrings split by the matching delimiters,
including the matching delimiters. public static String[] split(String s, String regex)
For example, split("ab#12#453", "#") returns ab, #, 12, #, and 453 in an array of String and split("a?b?gf#e", "[?#]") returns a, ?, b, ?, gf, #, and e in an array of String.