Create Exercise 05
(The Course class) Rewrite the Course class in Listing 10.6. Use an ArrayList to replace an array to store students. Draw the new UML diagram for the class.
You should not change the original contract of the Course class (i.e., the definition of the constructors and methods should not be changed, but the private
members may be changed.)