Create Exercise 19
Bin packing using first fit) The bin packing problem is to pack the objects of various weights into containers. Assume each container can hold a maximum of 10
pounds. The program uses an algorithm that places an object into the first bin in which it would fit. Your program should prompt the user to enter the total number
of objects and the weight of each object. The program displays the total number of containers needed to pack the objects and the contents of each container.
Here is a sample run of the program:
Enter the number of objects: 6
Enter the weights of the objects: 7 5 2 3 5 8
Container 1 contains objects with weight 7 2
Container 2 contains objects with weight 5 3
Container 3 contains objects with weight 5
Container 4 contains objects with weight 8
Does this program produce an optimal solution, that is, finding the minimum number of containers to pack the objects?