Create Chapter 12\Exercise 01
Listing 7.9, Calculator.java, is a simple command-linecalculator. Note the program terminates if any operand is
nonnumeric. Write a program with an exception handler that deals with nonnumeric operands; then write another program without using an exception
handler to achieve the same objective. Your program should display a message that informs the user of the wrong operand type before exiting (seeFigure 12.12).