Create Exercise 06
(NumberFormatException) Listing 6.8 implements the hex2Dec(StringhexString) method, which converts a hex string into a decimal number.
Implement the hex2Dec method to throw a NumberFormatException if the string is not a hex string. Write a test program that prompts the user to enter
a hex number as a string and displays its decimal equivalent. If the method throws an exception, display “Not a hex number”.