Create Exercise 12
Reformat Java source code) Write a program that converts the Java source code from the next-line brace style to the end-of-line brace style. For example,
the following Java source in (a) uses the next-line brace style. Your program converts it to the end-of-line brace style in (b).
public class Test
{
 public static void main(String[] args)
 {
 // Some statements
 }
}
(a) Next-line brace style

public class Test {
 public static void main(String[] args) {
 // Some statements
 }
}
(b) End-of-line brace style

Your program can be invoked from the command line with the Java source-code file as the argument. It converts the Java source code to a new format. For example,
the following command converts the Java source-code file Test.java to the end-of-line brace style. 
java Exercise12_12 Test.java