Create Exercise 13
(Count characters, words, and lines in a file) Write a program that will count the number of characters, words, and lines in a file. Words are separated by
whitespace characters. The file name should be passed as a command-line  argument, as shown in Figure 12.13.
The program displays the number of characters, words, and lines in the given file.